refactor: break down QuizGenerationPanel into smaller components
This refactoring improves maintainability and reusability by extracting focused components:

Components created:
- QuizGenerationCard: Encapsulates individual quiz card display logic (78 lines)
- QuizProgressIndicator: Reusable progress bar with phase text (30 lines)
- QuizBadges: Question count and LLM model badge display (26 lines)
- QuizGenerationPanelSkeleton: Dedicated skeleton loading component (60 lines)

Benefits:
- Reduced QuizGenerationPanel from 225+ lines to 82 lines (64% reduction)
- Added React.memo to components for performance optimization
- Improved testability with smaller, focused components
- Enhanced reusability across dashboard panels
- Better separation of concerns and single responsibility principle

Technical improvements:
- Fixed type safety for quiz.question_count with fallback
- Consistent use of UI constants for styling
- Proper TypeScript interfaces for all component props
- Maintained existing functionality while improving code structure
