+++ frontend/src/components/dashboard/index.ts
@@ -1,1 +1,8 @@
-export * from './panels'
+// Dashboard components
+export { QuizGenerationCard } from "./QuizGenerationCard"
+export { QuizProgressIndicator } from "./QuizProgressIndicator"
+export { QuizBadges } from "./QuizBadges"
+export { QuizGenerationPanelSkeleton } from "./QuizGenerationPanelSkeleton"
+
+// Dashboard panels
+export * from "./panels"
